# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Document Generation Dynamic API
  version: 1.0.0
  summary: Enterprise financial letter document generation API contract
  description: >-
    Formal v1 OpenAPI contract for template-based document generation. Template
    variables are constrained by each released template version contract.
servers:
  - url: /
    description: Deployment root. Paths include /api/{environment}/v1.
...
security:
  - ApiCredentialId: []
    ApiCredentialSecret: []
    AccessAccount: []
paths:
...
  /api/management/v1/security-audit/route-access-denied:
    post:
      tags: [Admin]
      operationId: reportRouteAccessDenied
      summary: Report a management forbidden-route denial
      description: >-
        Authenticated management clients call this after a client-side route
        guard denies access (LR-D7). Persists SECURITY_ROUTE_ACCESS_DENIED on
        management_audit_event using the session principal. Fail-safe: returns
        204 even when durable persistence fails. Does not accept passwords,
        tokens, or secrets. Unauthenticated callers receive 401.
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteAccessDeniedRequest'
      responses:
        '204':
          description: Denial recorded (or fail-safe ignored).
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    ApiCredentialId:
      type: apiKey
      in: header
      name: X-Api-Credential-Id
      description: Non-secret API credential identifier.
    ApiCredentialSecret:
      type: apiKey
      in: header
      name: X-Api-Credential-Secret
      description: API credential secret. It must not be logged, audited, returned, or displayed after submission.
    AccessAccount:
      type: apiKey
      in: header
      name: X-Access-Account
      description: User or service account whose AD Group membership is evaluated.
  parameters:
...
    TraceIdHeader:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 1
      description: Optional caller trace ID. If omitted, the platform generates traceId.
...
  responses:
...
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    RouteAccessDeniedRequest:
      type: object
      additionalProperties: false
      required: [routeKey, traceId]
      properties:
        routeKey:
          type: string
          minLength: 1
          maxLength: 256
          description: Logical management route key (preferred) or path summary.
          example: route.audit-console
        traceId:
          type: string
          minLength: 1
          maxLength: 128
          description: Client-side denial correlation id (no secrets).
    OutputFormat:
      type: string
      enum: [DOCX, PDF]
    OutputMode:
      type: string
      enum: [SYNC_STREAM, SYNC_DOWNLOAD_URL, ASYNC_TASK]
    RouteType:
      type: string
      enum: [EXPLICIT_VERSION, DEFAULT_ROUTE]
    FidelityWarningCode:
      type: string
      enum: [OPTIONAL_CONTENT_EMPTY, LOW_RISK_PAGINATION_DIFFERENCE, LOW_RISK_TABLE_PAGE_BREAK, CONTROLLED_STYLE_FALLBACK, IMAGE_SCALING_ADJUSTED]
    IdempotencyStatus:
      type: string
      enum: [IDEMPOTENCY_NEW, IDEMPOTENCY_REPLAYED, IDEMPOTENCY_CONFLICTED]
...
    BatchItemStatus:
      type: string
      enum: [SUCCEEDED, FAILED, SKIPPED]
    Permission:
      type: string
      enum: [ALLOW_PRINT, ALLOW_COPY, ALLOW_EDIT, ALLOW_ANNOTATE, ALLOW_FORM_FILL]
...
    OutputOptions:
      type: object
      additionalProperties: false
      required: [format, mode]
      properties:
        format:
          $ref: '#/components/schemas/OutputFormat'
        mode:
          $ref: '#/components/schemas/OutputMode'
...
    Metadata:
      type: object
      additionalProperties: false
      required: [auditId, traceId]
      properties:
        auditId:
          type: string
        traceId:
          type: string
        requestId:
          type: string
        idempotencyKey:
          type: string
          description: Echoed for generation success and generation error responses.
        idempotencyStatus:
          $ref: '#/components/schemas/IdempotencyStatus'
        originalRequestAt:
          type: string
          format: date-time
          description: Always returned for successful idempotent replay responses and for idempotency conflict safe summaries.
        templateId:
          type: string
        routeType:
          $ref: '#/components/schemas/RouteType'
        resolvedReleaseVersion:
          type: string
        output:
          $ref: '#/components/schemas/OutputOptions'
...
    DownloadInfo:
      type: object
      additionalProperties: false
      required: [url, expiresAt, requiresAuthorization, oneTime]
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
        requiresAuthorization:
          type: boolean
          const: true
        oneTime:
          type: boolean
          const: false
...
    BatchResultItem:
      type: object
      additionalProperties: false
      required: [itemId, status, output, encryptionSummary]
      properties:
        itemId:
          type: string
        status:
          $ref: '#/components/schemas/BatchItemStatus'
        output:
          $ref: '#/components/schemas/OutputOptions'
        encryptionSummary:
          $ref: '#/components/schemas/EncryptionSummary'
        documentId:
          type: string
        download:
          $ref: '#/components/schemas/DownloadInfo'
        error:
          $ref: '#/components/schemas/ErrorDetail'
        fidelityWarnings:
          type: array
          items:
            $ref: '#/components/schemas/FidelityWarning'
...
    FidelityWarning:
      type: object
      additionalProperties: false
      required: [warningCode, messageKey, message, locationSummary, detectedSummary, recommendation, sensitiveDataExcluded]
      properties:
        warningCode:
          $ref: '#/components/schemas/FidelityWarningCode'
        messageKey:
          type: string
        message:
          type: string
        locationSummary:
          type: string
        detectedSummary:
          type: string
        recommendation:
          type: string
        sensitiveDataExcluded:
          type: boolean
          const: true
    EncryptionSummary:
      type: object
      additionalProperties: false
      required: [enabled, outputFormat, openPasswordProvided, ownerPasswordProvided, permissions]
      properties:
        enabled:
          type: boolean
        outputFormat:
          $ref: '#/components/schemas/OutputFormat'
        openPasswordProvided:
          type: boolean
        ownerPasswordProvided:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [metadata, error]
      properties:
        metadata:
          $ref: '#/components/schemas/Metadata'
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, category, message, messageKey, retryable]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        category:
          $ref: '#/components/schemas/ErrorCategory'
        message:
          type: string
        messageKey:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        idempotencyConflict:
          type: object
          additionalProperties: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/BatchResultItem'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason, message]
      properties:
        field:
          type: string
        reason:
          type: string
          enum: [REQUIRED, INVALID_TYPE, INVALID_FORMAT, OUT_OF_RANGE, TOO_LONG, TOO_SHORT, ENUM_NOT_ALLOWED, PATTERN_MISMATCH, RULE_FAILED, DUPLICATED, UNKNOWN_FIELD]
        message:
          type: string
    ErrorCategory:
      type: string
      enum: [AUTHENTICATION, AUTHORIZATION, VERSION_ROUTING, API_POLICY, IDEMPOTENCY, VALIDATION, TEMPLATE_CONTRACT, GENERATION, ENCRYPTION, BATCH]
    ErrorCode:
      type: string
      enum:
        - API_CREDENTIAL_REQUIRED
        - API_CREDENTIAL_INVALID
        - API_CREDENTIAL_EXPIRED
        - API_CREDENTIAL_REVOKED
        - ACCESS_ACCOUNT_REQUIRED
        - AD_GROUP_RESOLUTION_FAILED
        - AD_GROUP_NOT_AUTHORIZED
        - TEMPLATE_ACCESS_DENIED
        - ENVIRONMENT_MISMATCH
        - RELEASE_VERSION_REQUIRED
        - RELEASE_VERSION_FORMAT_INVALID
        - RELEASE_VERSION_NOT_FOUND
        - RELEASE_VERSION_DISABLED
        - DEFAULT_ROUTE_NOT_CONFIGURED
        - DEFAULT_ROUTE_TARGET_UNAVAILABLE
        - TEMPLATE_DISABLED
        - TEMPLATE_DEPRECATED
        - OUTPUT_FORMAT_NOT_ALLOWED
        - OUTPUT_MODE_NOT_ALLOWED
        - BATCH_LIMIT_EXCEEDED
        - ENCRYPTION_NOT_ALLOWED
        - DOWNLOAD_URL_EXPIRED
        - RESULT_RETENTION_EXPIRED
        - IDEMPOTENCY_KEY_REQUIRED
        - IDEMPOTENCY_KEY_CONFLICT
        - IDEMPOTENCY_RETRY_NOT_ALLOWED
        - IDEMPOTENCY_STORE_UNAVAILABLE
        - REQUEST_BODY_INVALID
        - REQUEST_ID_REQUIRED
        - OUTPUT_FORMAT_REQUIRED
        - OUTPUT_MODE_REQUIRED
        - VARIABLES_REQUIRED
        - VARIABLE_REQUIRED
        - VARIABLE_TYPE_INVALID
        - VARIABLE_FORMAT_INVALID
        - VARIABLE_RULE_FAILED
        - TEMPLATE_CONTRACT_INVALID
        - TEMPLATE_ANCHOR_MISSING
        - DOCX_GENERATION_FAILED
        - PDF_CONVERSION_FAILED
        - GENERATION_TIMEOUT
        - GENERATION_SERVICE_UNAVAILABLE
        - ASYNC_TASK_NOT_FOUND
        - ASYNC_TASK_EXPIRED
        - ASYNC_TASK_CANCELLATION_NOT_ALLOWED
        - DOCUMENT_NOT_FOUND
        - WORK_ITEM_NOT_FOUND
        - ENCRYPTION_PARAMETER_INVALID
        - ENCRYPTION_FAILED
        - BATCH_ITEMS_REQUIRED
        - BATCH_ITEM_COUNT_INVALID
        - ITEM_ID_REQUIRED
        - ITEM_ID_DUPLICATED
        - BATCH_PARTIAL_FAILED
        - BATCH_PROCESSING_FAILED